Removing unused auth code, splitting some logic into its own testable unit.
