--- server/src/auth/auth.plugin.ts
@@ -1,35 +1,0 @@
-// Totally freestyling here
-import type {
-  ApolloServerPlugin,
-  GraphQLRequestListener,
-} from "@apollo/server";
-import { GraphQLError } from "graphql";
-import type { GraphQLContext } from "./auth.util";
-
-const PUBLIC_FIELDS = new Set<string>([
-  "Query.__schema",
-  "Query.__type",
-  "Query.health",
-  "Query.loginStatus",
-]);
-
-export const authGatePlugin: ApolloServerPlugin<GraphQLContext> = {
-  async requestDidStart(): Promise<GraphQLRequestListener<GraphQLContext>> {
-    return {
-      async executionDidStart() {
-        return {
-          willResolveField(fieldResolverParams) {
-            const { info, contextValue } = fieldResolverParams;
-            const path = `${info.parentType.name}.${info.fieldName}`;
-
-            if (!contextValue?.user && !PUBLIC_FIELDS.has(path)) {
-              throw new GraphQLError("User not authenticated", {
-                extensions: { code: "UNAUTHENTICATED", http: { status: 401 } },
-              });
-            }
-          },
-        };
-      },
-    };
-  },
-};
