+++ server/src/local-server.ts
@@ -9,7 +9,6 @@ import {
   type GraphQLContext,
   validateClaims,
 } from "./auth/auth.util.js";
-import { authGatePlugin } from "./auth/auth.plugin.js";
 import { gatedLandingPagePlugin } from "./plugins/gatedLandingPage.plugin.js";
 import { als, log, reqIdChild, store } from "./log.js";
 import { loggingPlugin } from "./plugins/logging.plugin";
@@ -17,6 +16,7 @@ import { GraphQLArmorConfig } from "./plugins/graphQLArmorConfig.js";
 import { JwtPayload } from "jsonwebtoken";
 import { parseCookie } from "cookie";
 import { decodeToken } from "./auth/decodeToken.js";
+import { getPersonTypeFromClaims } from "./auth/getPersonTypeFromClaims.js";
 import { fieldAuthPlugin } from "./plugins/fieldAuthPlugin.js";
 import { formatGraphQLErrorCode } from "./errors/errorCodes.js";
 
@@ -30,13 +30,7 @@ const server = new ApolloServer<GraphQLContext>({
   resolvers,
   ...protection,
   introspection: process.env.ALLOW_INTROSPECTION === "true",
-  plugins: [
-    ...protection.plugins,
-    authGatePlugin,
-    loggingPlugin,
-    fieldAuthPlugin,
-    gatedLandingPagePlugin(),
-  ],
+  plugins: [...protection.plugins, loggingPlugin, fieldAuthPlugin, gatedLandingPagePlugin()],
   validationRules: [...protection.validationRules],
   formatError: formatGraphQLErrorCode,
   logger: log,
@@ -69,6 +63,9 @@ const { url } = await startStandaloneServer<GraphQLContext>(server, {
 
       const decodedToken = await decodeToken(token);
       const claims = extractClaimsFromDecodedToken(decodedToken);
+      // Validate that the claims map to a valid person type (role gate);
+      // throws and rejects the request if the role is missing or ambiguous.
+      getPersonTypeFromClaims(claims);
       const ctx = await buildContextFromClaims(claims);
 
       const requestId = (req.headers["x-request-id"] as string | undefined) || randomUUID();
