+++ server/src/server.test.ts
@@ -25,7 +25,6 @@ vi.mock("@as-integrations/aws-lambda", () => ({
 vi.mock("@escape.tech/graphql-armor", () => ({ ApolloArmor: ApolloArmorMock }));
 
 vi.mock("./model/graphql.js", () => ({ typeDefs: [], resolvers: {} }));
-vi.mock("./auth/auth.plugin.js", () => ({ authGatePlugin: {} }));
 vi.mock("./plugins/logging.plugin", () => ({ loggingPlugin: {} }));
 vi.mock("./plugins/graphQLArmorConfig.js", () => ({ GraphQLArmorConfig: { __sentinel: true } }));
 vi.mock("./auth/auth.util.js", () => ({
